cyberframework/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-WTF==1.2.2
WTForms==3.2.1
weasyprint==63.1
PyJWT==2.10.1
cryptography==44.0.0
python-dotenv==1.0.1
requests==2.32.3
email-validator==2.2.0
openpyxl==3.1.5
